refactor: renaming, use proper terminology.
config_facade.ConfigFacade was not a "Facade" that
facilitates access to some subcomponents, but
rather functions as a "Adapter" that allows the
non-refactored code to interact with the refactored
configuration.

Accordingly it was renamed to config_adapter.Config
and all references to a "Facade" in the documentation/
test function names where updated as well.